+++ Runtime/Unity/Util/Editor/ReadOnlyAttribute.cs
@@ -1,0 +1,34 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2016-2021 Juan Delgado (@JuDelCo)
+
+#if UNITY_2019_3_OR_NEWER
+#if UNITY_EDITOR
+
+using UnityEngine;
+using UnityEditor;
+
+namespace Ju.Util
+{
+	public class ReadOnlyAttribute : PropertyAttribute
+	{
+	}
+
+	[CustomPropertyDrawer(typeof(ReadOnlyAttribute))]
+	public class ReadOnlyDrawer : PropertyDrawer
+	{
+		public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
+		{
+			return EditorGUI.GetPropertyHeight(property, label, true);
+		}
+
+		public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+		{
+			GUI.enabled = false;
+			EditorGUI.PropertyField(position, property, label, true);
+			GUI.enabled = true;
+		}
+	}
+}
+
+#endif
+#endif